Refactor: eventHandler객체 지우기
