+++ src/layout/Header/Category/eventHandler.js
@@ -1,0 +1,16 @@
+import {
+  findTargetIdElement,
+  handleDisplayElement,
+} from '../../../utils/manuplateDOM.js';
+
+export function handleClick({ target }) {
+  const $categoryButton = target.closest('.category__button');
+  if ($categoryButton) {
+    handleCategoryButton.call(this);
+  }
+}
+
+function handleCategoryButton() {
+  const $categoryLayer = findTargetIdElement(this.$element, 'category-layer');
+  handleDisplayElement($categoryLayer);
+}
